requirements.txt
kafka-python==2.0.2
lz4==4.0.1
